Feature/make db async

+++ backend/core/app/models/configuration.py
@@ -2,7 +2,8 @@ from sqlalchemy import Boolean, Column, ForeignKey, Integer, String
 from sqlalchemy.types import BLOB
 from sqlalchemy.orm import relationship, Session
 from .ids_container import IdsContainer
-
+from sqlalchemy.future import select
+from sqlalchemy.ext.asyncio import AsyncSession
 from ..database import Base
 
 class Configuration(Base):
@@ -14,24 +15,31 @@ class Configuration(Base):
     file_type = Column(String(32), nullable=False)
     description = Column(String(2048))
 
-    container = relationship("IdsContainer", back_populates="configuration", foreign_keys=[IdsContainer.configuration_id])
-    containerRuleset = relationship('IdsContainer', back_populates='ruleset', foreign_keys=[IdsContainer.ruleset_id])
+    container = relationship("IdsContainer", foreign_keys=[IdsContainer.configuration_id])
+    containerRuleset = relationship('IdsContainer', foreign_keys=[IdsContainer.ruleset_id])
 
-def get_config_by_id(db: Session, config_id: int):
-    return db.query(Configuration).filter(Configuration.id == config_id).first()
-    
-def get_all_configurations(db: Session):
-    return db.query(Configuration).all()
+async def get_config_by_id(db: AsyncSession, config_id: int):
+    stmt = select(Configuration).where(Configuration.id == config_id)
+    result = await db.execute(stmt)
+    return result.scalar_one_or_none() 
 
-def remove_configuration_by_id(db, id):
-    config = get_config_by_id(db, id)
-    db.delete(config)
-    db.commit()
+async def get_all_configurations(db: AsyncSession):
+    stmt = select(Configuration)
+    result = await db.execute(stmt)
+    return result.scalars().all()  
 
-def add_config(db: Session, configuration: Configuration):
+async def remove_configuration_by_id(db: AsyncSession, config_id: int):
+    config = await get_config_by_id(db, config_id)
+    if config:
+        await db.delete(config)
+        await db.commit()
+
+async def add_config(db: AsyncSession, configuration: Configuration):
     db.add(configuration)
-    db.commit()
+    await db.commit() 
+    await db.refresh(configuration)  
 
-
-def get_all_configurations_by_type(db: Session, file_type: str):
-    return db.query(Configuration).filter(Configuration.file_type == file_type).all()
+async def get_all_configurations_by_type(db: AsyncSession, file_type: str):
+        stmt = select(Configuration).where(Configuration.file_type == file_type)
+        result = await db.execute(stmt)
+        return result.scalars().all()
